Break up code into components for clarity
