+++ src/components/MovieInfo.js
@@ -1,0 +1,12 @@
+import MovieDetail from "../components/MovieDetail";
+import MovieVideos from "../components/MovieVideos";
+import "./MovieInfo.css";
+
+export default function MovieInfo({ movie }) {
+  return (
+    <div className="movie-info--container">
+      <MovieDetail movie={movie} />
+      <MovieVideos movieId={movie.id} />
+    </div>
+  );
+}
