+++ src/components/MovieVideos.css
@@ -1,0 +1,4 @@
+.react-player-container {
+  width: 100%;
+  aspect-ratio: 16 / 9;
+}
