+++ src/pages/MoviePage.css
@@ -1,33 +1,3 @@
-.movie-details--container {
-  padding: 150px 20px 20px;
-  display: grid;
-  grid-template-columns: 30% 70%;
-
-  gap: 10px;
-
-  /* reset the CSS margins and fix rendering issues */
-  * {
-    z-index: 1;
-    margin: 0;
-  }
-}
-
-.movie-details {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-
-  img {
-    width: 100%;
-    max-width: 150px;
-  }
-
-  .movie-info td:first-child {
-    width: 120px;
-    color: #8d8d8d;
-  }
-}
-
 .return-home-button {
   position: absolute;
   top: 80px;
@@ -51,9 +21,4 @@
     border-color: #73cb3e;
     background-color: #73cb3e;
   }
-}
-
-.react-player-container {
-  width: 100%;
-  aspect-ratio: 16 / 9;
 }
